Decouple participantsService from the store
Moved joinConversation, leaveConversation and a few other dependencies
to the participantsStore.
Removes the dependency to the store from participantsService.

There's a slight change of behavior: joinConversation in the service now
also connects to signaling, and only when this is done the first
action dispatch are done. In the previous implementation the store
already received some updates before we even connected to signaling.
