+++ src/components/LeftSidebar/ConversationsList/ConversationsList.vue
@@ -40,7 +40,6 @@
 import Conversation from './Conversation'
 import Hint from '../../Hint'
 import LoadingPlaceholder from '../../LoadingPlaceholder'
-import { joinConversation, leaveConversation } from '../../../services/participantsService'
 import { EventBus } from '../../../services/EventBus'
 
 export default {
@@ -111,10 +110,10 @@ export default {
 				return
 			}
 			if (from.name === 'conversation') {
-				leaveConversation(from.params.token)
+				this.$store.dispatch('leaveConversation', { token: from.params.token })
 			}
 			if (to.name === 'conversation') {
-				joinConversation(to.params.token)
+				this.$store.dispatch('joinConversation', { token: to.params.token })
 			}
 		},
